Raise bing errors for statuses 403, 500, 503
Bing returns the status codes within the json response. When a request is
forbidden, it will return a 403. If there is an internal server error,
it will return a 500. And if the service is unavailable for some other
reason, it will return a 503. That information was obtained from bing's
[Status Codes and Error
Handling](https://msdn.microsoft.com/en-us/library/ff701703.aspx)
documentation.

We handle those three status codes here to raise
`Geocoder::RequestDenied` (for 403) and `Geocoder::ServiceUnavailable`
(for 500 and 503). This allows for the users of the Geocoder to handle
those errors in their applications.